+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added `stopRefresh` function
 - Added `getDataSource` function
 - Added `donut` option to pie chart
+- Added `eachChart` function
 - Remove colors from tooltips for Chart.js pie chart
 
 ## 2.1.2